Power-up changes display above items. Other small fixes and changes.

+++ src/main/java/com/github/jmatcj/ld43/util/Util.java
@@ -56,7 +56,7 @@ public class Util {
      * @param y The Y location to start drawing at.
      */
     // TODO: Fix this with whatever font we're going to use
-    public static void drawText(GraphicsContext gc, Color fontColor, int fontSize, TextAlignment alignment, String text, int x, int y) {
+    public static void drawText(GraphicsContext gc, Color fontColor, int fontSize, TextAlignment alignment, String text, double x, double y) {
         gc.setFill(fontColor);
         gc.setFont(Font.font("Verdana", FontWeight.BOLD, fontSize));
         //gc.setFont(Font.loadFont(AssetLoader.getFontLoc(), fontSize));
